Deprecated functions in package semoutput.

The functions listed below are deprecated and will be defunct in the near future. When possible, alternative functions with similar functionality are also mentioned. Help pages for deprecated functions are available at help("<function>-deprecated").

sem.descriptives(x)

sem.factorcor(x, standardized = TRUE, factors = c(), print = TRUE)

sem.factorloadings(x, standardized = TRUE, print = TRUE)

sem.factorvar(x, factors = c(), standardized = TRUE, print = TRUE)

sem.fitmeasures(x, print = TRUE)

sem.paths(x, standardized = TRUE, print = TRUE)

sem.residuals(x, print = TRUE)

sem.rsquared(x, print = TRUE)

sem.sig(x, print = TRUE)

sem.descriptives

For sem.descriptives, use sem_descriptives.

sem.factorcor

For sem.factorcor, use sem_factorcor.

sem.factorloadings

For sem.factorloadings, use sem_factorloadings.

sem.factorvar

For sem.factorvar, use sem_factorvar.

sem.fitmeasures

For sem.fitmeasures, use sem_fitmeasures.

sem.paths

For sem.paths, use sem_paths.

sem.residuals

For sem.residuals, use sem_residuals.

sem.rsquared

For sem.rsquared, use sem_rsquared.

sem.sig

For sem.sig, use sem_sig.

Examples

sem.descriptives(x)

#> Warning: 'sem.descriptives' is deprecated.
#> Use 'sem_descriptives' instead.
#> See help("Deprecated")

#> Error in tidyr::gather(x, "Variable", "value"): object 'x' not found

sem.factorcor(x)

#> Warning: 'sem.factorcor' is deprecated.
#> Use 'sem_factorcor' instead.
#> See help("Deprecated")

#> Error in "lavaan.fsr" %in% class(object): object 'x' not found

sem.factorloadings(x)

#> Warning: 'sem.factorloadings' is deprecated.
#> Use 'sem_factorloadings' instead.
#> See help("Deprecated")

#> Error in "lavaan.fsr" %in% class(object): object 'x' not found

sem.factorvar(x)

#> Warning: 'sem.factorvar' is deprecated.
#> Use 'sem_factorvar' instead.
#> See help("Deprecated")

#> Error in "lavaan.fsr" %in% class(object): object 'x' not found

sem.fitmeasures(x)

#> Warning: 'sem.fitmeasures' is deprecated.
#> Use 'sem_fitmeasures' instead.
#> See help("Deprecated")

#> Error in lavaan::fitMeasures(x, c("cfi", "rmsea", "rmsea.ci.lower", "rmsea.ci.upper",     "aic", "bic")): object 'x' not found

sem.paths(x)

#> Warning: 'sem.paths' is deprecated.
#> Use 'sem_paths' instead.
#> See help("Deprecated")

#> Error in "lavaan.fsr" %in% class(object): object 'x' not found

sem.residuals(x)

#> Warning: 'sem.residuals' is deprecated.
#> Use 'sem_residuals' instead.
#> See help("Deprecated")

#> Error in lavaan::residuals(x, type = "cor"): object 'x' not found

sem.rsquared(x)

#> Warning: 'sem.rsquared' is deprecated.
#> Use 'sem_rsquared' instead.
#> See help("Deprecated")

#> Error in lavaan::inspect(x, "r2"): object 'x' not found

sem.sig(x)

#> Warning: 'sem.sig' is deprecated.
#> Use 'sem_sig' instead.
#> See help("Deprecated")

#> Error in lavaan::fitMeasures(x, c("ntotal", "chisq", "pvalue", "df")): object 'x' not found